Settings: Remove search button from subsettings
Showing the search button as part of the action bar in subsettings
screens creates a confusing hierarchy where it looks like the button is
meant for searching in the subsettings page, but it actually leads to
global Settings search. It also necessitates an alternative icon that
is somewhat confusing (find in page) for pages that actually do have
intra-page search.

Change-Id: I4936ac8a4b4f47b45f6e3060ce47dcf9d5bb33de